Move asymmetric sync example out of experimental namespace
